requirements.txt
asgiref==3.7.2
colorama==0.4.6
Django==4.2.9
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
iniconfig==2.0.0
mysqlclient==2.2.4
packaging==24.1
pluggy==1.5.0
PyJWT==2.9.0
pytest==8.3.3
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
tzdata==2024.1